warn users if quantization using insufficient nsamples
